+++ checkpoint/CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Roll back earlier change altering metadata format, which was observed to cause
 breakages.
+- Fix `CompositeHandler` extraction logic for handler type strings when missing `_CHECKPOINT_METADATA`
+
+### Changed
+
+- Split v0 checkpoint format/layout logic out from `OrbaxLayout` and into
+seperate `V0Layout`.
 
 ## [0.11.29] - 2025-11-25