+++ checkpoint/CHANGELOG.md
@@ -9,10 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.30] - 2025-11-26
 
+### Added
+
+- Added testing to verify backwards compatibility when a checkpoint is missing
+`_CHECKPOINT_METADATA` or saving pytree directly as a checkpoint with no
+checkpointable.
+
 ### Fixed
 
 - Roll back earlier change altering metadata format, which was observed to cause
 breakages.
+- Fix `CompositeHandler` extraction logic for handler type strings when missing `_CHECKPOINT_METADATA`
+
+### Changed
+
+- Split v0 checkpoint format/layout logic out from `OrbaxLayout` and into
+seperate `V0Layout`.
 
 ## [0.11.29] - 2025-11-25